// packages/frontend/src/generated/types/CompanyResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `company` entity.
 */
export interface CompanyResponse {
  /**
   * @type string
   */
  commonName: string
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type boolean
   */
  isContractor: boolean
  /**
   * @type boolean
   */
  isExporter: boolean
  /**
   * @type boolean
   */
  isManufacturer: boolean
  /**
   * @type boolean
   */
  isSender: boolean
  /**
   * @type string,null
   */
  legalName?: string | null
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}